Library

Documentation for SatelliteToolboxCelestialBodies.jl.

SatelliteToolboxCelestialBodies.SatelliteToolboxCelestialBodies — Module

module SatelliteToolboxCelestialBodies

Compute the position and velocity of celestial bodies (Sun and Moon) for the SatelliteToolbox.jl ecosystem.

SatelliteToolboxCelestialBodies._TAB_47A — Constant

const _TAB_47A::NTuple{60, SVector{6, Int32}}

Periodic terms for the longitude (Σl) and distance (Σr) of the Moon, obtained from the table 47.A in [1, pp. 339-340].

Each element contains, in order, the multipliers of the fundamental arguments D, M, M´, and F [-], the coefficient of the sine term for the longitude [10⁻⁶ °], and the coefficient of the cosine term for the distance [m].

References

[1] Meeus, J. (1998). Astronomical Algorithms. 2nd ed. Willmann-Bell, Inc, Richmond, VA.

SatelliteToolboxCelestialBodies._TAB_47B — Constant

const _TAB_47B::NTuple{60, SVector{5, Int32}}

Periodic terms for the latitude (Σb) of the Moon, obtained from the table 47.B in [1, p. 341].

Each element contains, in order, the multipliers of the fundamental arguments D, M, M´, and F [-], and the coefficient of the sine term for the latitude [10⁻⁶ °].

References

[1] Meeus, J. (1998). Astronomical Algorithms. 2nd ed. Willmann-Bell, Inc, Richmond, VA.

SatelliteToolboxCelestialBodies._evalpoly_with_derivative — Method

_evalpoly_with_derivative(x::Number, coefficients::NTuple{N, Number}) -> Number, Number

Evaluate the polynomial with the coefficients (c₀, c₁, ..., cₙ), ordered from the lowest to the highest degree, and its first derivative at x, both computed with the Horner scheme. The tuple coefficients must have at least two elements.

Returns

Number: Polynomial value c₀ + c₁ x + ... + cₙ xⁿ.

Number: Polynomial derivative c₁ + 2 c₂ x + ... + n cₙ xⁿ⁻¹.

SatelliteToolboxCelestialBodies._periodic_terms_table — Method

_periodic_terms_table(::Val{:Table47A}) -> NTuple{60, SVector{6, Int32}}
_periodic_terms_table(::Val{:Table47B}) -> NTuple{60, SVector{5, Int32}}

Return the table of periodic terms selected by the tag: _TAB_47A for Val(:Table47A) and _TAB_47B for Val(:Table47B).

SatelliteToolboxCelestialBodies._sincos_add — Method

_sincos_add(sc_a::Tuple{Number, Number}, sc_b::Tuple{Number, Number}) -> Number, Number

Compute the sine and cosine of the sum of two angles given the tuples sc_a and sc_b containing the sine and cosine of each angle, using the angle addition formulas.

Returns

Number: Sine of the sum of the angles.

Number: Cosine of the sum of the angles.

SatelliteToolboxCelestialBodies._sincos_multiple — Method

_sincos_multiple(sc_multiples::Tuple{Tuple{T, T}, Vararg{Tuple{T, T}}}, k::Integer) -> T, T

Return the sine and cosine of kθ given the tuple sc_multiples with the sines and cosines of the non-negative multiples of θ, as computed by _sincos_multiples. The multiplier k can be negative, in which case the parity of the sine and cosine functions is used. abs(k) must be lower than the length of sc_multiples.

Returns

T: Sine of kθ.

T: Cosine of kθ.

SatelliteToolboxCelestialBodies._sincos_multiples — Method

_sincos_multiples(sc_θ::Tuple{T, T}, ::Val{N}) -> NTuple{N + 1, Tuple{T, T}}

Compute the sine and cosine of the integer multiples 0θ, 1θ, ..., Nθ of an angle given the tuple sc_θ with its sine and cosine, using the angle addition formulas.

The result is a tuple in which the element k + 1 contains the sine and cosine of kθ.

Returns

NTuple{N + 1, Tuple{T, T}}: Sines and cosines of the multiples of the angle.

SatelliteToolboxCelestialBodies._sincos_sub — Method

_sincos_sub(sc_a::Tuple{Number, Number}, sc_b::Tuple{Number, Number}) -> Number, Number

Compute the sine and cosine of the difference of two angles given the tuples sc_a and sc_b containing the sine and cosine of the minuend and subtrahend angles, using the angle subtraction formulas.

Returns

Number: Sine of the difference of the angles.

Number: Cosine of the difference of the angles.

SatelliteToolboxCelestialBodies._sum_periodic_terms — Method

_sum_periodic_terms(
    ::Val{table},
    sc_D::Tuple{T, T},
    sc_M::Tuple{T, T},
    sc_M´::Tuple{T, T},
    sc_F::Tuple{T, T},
    ∂D::Number,
    ∂M::Number,
    ∂M´::Number,
    ∂F::Number,
    E::Number,
    ∂E::Number
) -> Number, Number, Number, Number

Sum the periodic terms of the table, which must be :Table47A [2, pp. 339-340] or :Table47B [2, p. 341], given the tuples sc_D, sc_M, sc_M´, and sc_F with the sines and cosines of the fundamental arguments D, M, M´, and F, their time derivatives ∂D, ∂M, ∂M´, and ∂F [rad/century], the eccentricity correction factor E [-], and its time derivative ∂E [1/century].

The sums are computed as:

\[Σ_s = \sum_k c_{s, k} E_k \sin(θ_k), \quad Σ_c = \sum_k c_{c, k} E_k \cos(θ_k),\]

where θ_k is the integer combination of the fundamental arguments of the term k, c_{s, k} and c_{c, k} are its sine and cosine coefficients, and E_k is E, E², or 1 depending on the multiplier of M. The tables without cosine coefficients yield zero cosine sums.

The function is generated from the selected table so that the multipliers become literal constants: the sines and cosines of the arguments are composed from those of the fundamental arguments using the angle addition formulas, and the terms with null multipliers are removed at compile time. Hence, the sines and cosines in sc_D, sc_M, sc_M´, and sc_F must be accurate since every term is derived from them.

Returns

Number: Sum of the sine terms Σₛ, in the unit of the sine coefficients of the table.

Number: Sum of the cosine terms Σ_c, in the unit of the cosine coefficients of the table.

Number: Time derivative of Σₛ [1/century].

Number: Time derivative of Σ_c [1/century].

References

[2] Meeus, J. (1998). Astronomical Algorithms. 2nd ed. Willmann-Bell, Inc, Richmond, VA.

SatelliteToolboxCelestialBodies.moon_position_mod — Method

moon_position_mod(jd_tdb::Number[, model]) -> SVector{3, T}
moon_position_mod(date_tdb::Union{Date, DateTime}[, model]) -> SVector{3, Float64}

Compute the Moon position [m] represented in the IAU-76/FK5 MOD (mean-equator, mean-equinox of date) at the Julian Day jd_tdb or at the date date_tdb, both in the Barycentric Dynamical Time (TDB).

The model selects the algorithm and must be Val(:Meeus) [2, ch. 47] or Val(:Vallado) [1, p. 288]. See moon_state_mod for the details on the accuracy of each model and on the element type T of the result. The function throws an error if model is not one of the supported models.

See also: moon_state_mod, moon_velocity_mod

Arguments

jd_tdb::Number: Julian Day [TDB] at which the position must be computed.

date_tdb::Union{Date, DateTime}: Date [TDB] at which the position must be computed.

model::Val: Algorithm used to compute the Moon position. (Default: Val(:Meeus))

Returns

SVector{3, T}: Moon position vector [m] represented in MOD.

References

[1] Vallado, D. A. (2013). Fundamentals of Astrodynamics and Applications. 4th ed. Microcosm Press, Hawthorne, CA.

[2] Meeus, J. (1998). Astronomical Algorithms. 2nd ed. Willmann-Bell, Inc, Richmond, VA.

Extended help

Throws

ArgumentError: model is neither Val(:Meeus) nor Val(:Vallado).

SatelliteToolboxCelestialBodies.moon_state_mod — Method

moon_state_mod(jd_tdb::Number[, model]) -> SVector{3, T}, SVector{3, T}
moon_state_mod(
    date_tdb::Union{Date, DateTime}[, model]
) -> SVector{3, Float64}, SVector{3, Float64}

Compute the Moon position [m] and velocity [m/s] represented in the IAU-76/FK5 MOD (mean-equator, mean-equinox of date) at the Julian Day jd_tdb or at the date date_tdb, both in the Barycentric Dynamical Time (TDB).

The model selects the algorithm and must be Val(:Meeus) or Val(:Vallado). Val(:Meeus) uses the algorithm in [2, ch. 47] that provides an accuracy of 10 [arcsec] in the longitude and 4 [arcsec] in the latitude (the reference does not mention the timespan). Val(:Vallado) uses the algorithm in [1, p. 288] that is about 3 times faster than Val(:Meeus) but can lead to errors of 0.3 [°] in longitude and 0.2 [°] in latitude. In both cases, the velocity is the analytical time derivative of the position. Prefer this function over calling moon_position_mod and moon_velocity_mod separately when both quantities are required, since the computation is shared.

The function throws an error if model is not one of the supported models.

See also: moon_position_mod, moon_velocity_mod

Arguments

jd_tdb::Number: Julian Day [TDB] at which the state must be computed.

date_tdb::Union{Date, DateTime}: Date [TDB] at which the state must be computed.

model::Val: Algorithm used to compute the Moon state. (Default: Val(:Meeus))

Returns

SVector{3, T}: Moon position vector [m] represented in MOD.

SVector{3, T}: Moon velocity vector [m/s] represented in MOD.

References

[1] Vallado, D. A. (2013). Fundamentals of Astrodynamics and Applications. 4th ed. Microcosm Press, Hawthorne, CA.

[2] Meeus, J. (1998). Astronomical Algorithms. 2nd ed. Willmann-Bell, Inc, Richmond, VA.

Extended help

The element type T of the result is promote_type(float(typeof(jd_tdb)), Float64). Hence, Float32 inputs yield Float64 results because the precision of the series requires it, whereas wider types, such as BigFloat or automatic differentiation numbers, propagate to the output.

Throws

ArgumentError: model is neither Val(:Meeus) nor Val(:Vallado).

SatelliteToolboxCelestialBodies.moon_velocity_mod — Method

moon_velocity_mod(jd_tdb::Number[, model]) -> SVector{3, T}
moon_velocity_mod(date_tdb::Union{Date, DateTime}[, model]) -> SVector{3, Float64}

Compute the Moon velocity [m/s] measured and represented in the IAU-76/FK5 MOD (mean-equator, mean-equinox of date) at the Julian Day jd_tdb or at the date date_tdb, both in the Barycentric Dynamical Time (TDB).

The model selects the algorithm and must be Val(:Meeus) [2, ch. 47] or Val(:Vallado) [1, p. 288]. In both cases, the velocity is the analytical time derivative of the Moon position. See moon_state_mod for the details on the accuracy of each model and on the element type T of the result. The function throws an error if model is not one of the supported models.

See also: moon_state_mod, moon_position_mod

Arguments

jd_tdb::Number: Julian Day [TDB] at which the velocity must be computed.

date_tdb::Union{Date, DateTime}: Date [TDB] at which the velocity must be computed.

model::Val: Algorithm used to compute the Moon velocity. (Default: Val(:Meeus))

Returns

SVector{3, T}: Moon velocity vector [m/s] represented in MOD.

References

[1] Vallado, D. A. (2013). Fundamentals of Astrodynamics and Applications. 4th ed. Microcosm Press, Hawthorne, CA.

[2] Meeus, J. (1998). Astronomical Algorithms. 2nd ed. Willmann-Bell, Inc, Richmond, VA.

Extended help

Throws

ArgumentError: model is neither Val(:Meeus) nor Val(:Vallado).

SatelliteToolboxCelestialBodies.sun_position_mod — Method

sun_position_mod(jd_tdb::Number) -> SVector{3, T}
sun_position_mod(date_tdb::Union{Date, DateTime}) -> SVector{3, Float64}

Compute the Sun position [m] represented in the IAU-76/FK5 MOD (mean-equator, mean-equinox of date) at the Julian Day jd_tdb or at the date date_tdb, both in the Barycentric Dynamical Time (TDB).

The algorithm is described in [1, pp. 277-279]. See sun_state_mod for the details on the accuracy and on the element type T of the result.

See also: sun_state_mod, sun_velocity_mod

Returns

SVector{3, T}: Sun position vector [m] represented in MOD.

References

[1] Vallado, D. A. (2013). Fundamentals of Astrodynamics and Applications. 4th ed. Microcosm Press, Hawthorne, CA.

SatelliteToolboxCelestialBodies.sun_state_mod — Method

sun_state_mod(jd_tdb::Number) -> SVector{3, T}, SVector{3, T}
sun_state_mod(date_tdb::Union{Date, DateTime}) -> SVector{3, Float64}, SVector{3, Float64}

Compute the Sun position [m] and velocity [m/s] represented in the IAU-76/FK5 MOD (mean-equator, mean-equinox of date) at the Julian Day jd_tdb or at the date date_tdb, both in the Barycentric Dynamical Time (TDB).

The position follows the algorithm in [1, pp. 277-279] and the velocity is its analytical time derivative. Prefer this function over calling sun_position_mod and sun_velocity_mod separately when both quantities are required, since the computation is shared.

See also: sun_position_mod, sun_velocity_mod

Returns

SVector{3, T}: Sun position vector [m] represented in MOD.

SVector{3, T}: Sun velocity vector [m/s] represented in MOD.

References

[1] Vallado, D. A. (2013). Fundamentals of Astrodynamics and Applications. 4th ed. Microcosm Press, Hawthorne, CA.

Extended help

The element type T of the result is promote_type(float(typeof(jd_tdb)), Float64). Hence, Float32 inputs yield Float64 results because the precision of the series requires it, whereas wider types, such as BigFloat or automatic differentiation numbers, propagate to the output.

The algorithm uses the number of Julian centuries in TDB for all fundamental arguments, including the mean longitude of the Sun, which is defined in UT1 in [1]. The resulting error is below 1'' and negligible for the accuracy of this model.

SatelliteToolboxCelestialBodies.sun_velocity_mod — Method

sun_velocity_mod(jd_tdb::Number) -> SVector{3, T}
sun_velocity_mod(date_tdb::Union{Date, DateTime}) -> SVector{3, Float64}

Compute the Sun velocity [m/s] measured and represented in the IAU-76/FK5 MOD (mean-equator, mean-equinox of date) at the Julian Day jd_tdb or at the date date_tdb, both in the Barycentric Dynamical Time (TDB).

The velocity is the analytical time derivative of the Sun position described in [1, pp. 277-279]. See sun_state_mod for the details on the accuracy and on the element type T of the result.

See also: sun_state_mod, sun_position_mod

Returns

SVector{3, T}: Sun velocity vector [m/s] represented in MOD.

References

[1] Vallado, D. A. (2013). Fundamentals of Astrodynamics and Applications. 4th ed. Microcosm Press, Hawthorne, CA.